MRADEV-943 Migrate Database
Moved the code to create the databse out to a start script that is run when the application starts. This removes the unnecessary processing overhead from the applicaiton that checks if the table is created on every call of the get_users_table() function. Because it is not the responsibility of the application to check if the database is configured correctly, this is more inline with best practices. This also removes extraneous logging from the mra logs, making it easier to understand and debug (previously each user insert would cause "errors" that could confuse developers unfamiliar with the issue).